+++ app.py
@@ -1,18 +1,19 @@
 from flask import Flask, render_template, jsonify, request
 import flask
 import time
-import yaml
 import re
 from classversion import ocr_crawler
 import threading
 from pathlib import Path
-from datetime import datetime
+from views.ymls import ymlroute
 import util
 
 craw: ocr_crawler = None
 app = Flask(__name__, static_folder='src', static_url_path='/src')
+app.register_blueprint(ymlroute)
 home = Path.cwd()
-shops = [file.name for file in (home / 'cite_envs').iterdir() if file.is_file()]
+shops = [file.name for file in (
+    home / 'cite_envs').iterdir() if file.is_file()]
 
 
 def background_task(mode):
@@ -46,7 +47,6 @@ def start_cite():
     return jsonify(redirect_url='/src/index.html')
 
 
-
 @app.route('/run_test')
 def run_test():
     thread = threading.Thread(target=background_task, args=('test',))
@@ -59,50 +59,6 @@ def run_all():
     thread = threading.Thread(target=background_task, args=('all',))
     thread.start()
     return jsonify(message="完整執行開始")
-
-
-
-@app.route('/get_yml', methods=['GET'])
-def get_yml():
-    global shops
-    base_path = home / 'cite_envs'
-    shop = util.remove_non_alphanumeric(request.args.get('shop'))
-    shop += '.yml'
-    fullpath = base_path / shop
-    if base_path not in fullpath.parents:
-        return jsonify({"message": "Not allowed"}), 400
-    if shop not in shops:
-        return jsonify({"message": "Invalid data"}), 400
-    with open(fullpath, 'r') as file:
-        cite_config = yaml.safe_load(file)
-
-    # 其他的代码...
-    return jsonify(cite_config)
-
-
-@app.route('/save_yml/<shop>', methods=['POST'])
-def save_yml(shop):
-    global shops
-    shop = util.remove_non_alphanumeric(shop)
-    data_to_save = request.json
-
-    # 检查data_to_save是否为有效数据
-    if not data_to_save:
-        return jsonify({"message": "Invalid data"}), 400
-
-    base_path = home / 'cite_envs'
-    shop = shop + '.yml'
-    fullpath = base_path / shop
-    if base_path not in fullpath.parents:
-        return jsonify({"message": "Not allowed"}), 400
-    if shop not in shops:
-        return jsonify({"message": "Invalid data"}), 400
-    data_to_save['lastupdate'] = str(datetime.now())
-    # 写入文件
-    with open(fullpath, 'w') as file:
-        yaml.safe_dump(data_to_save, file)
-
-    return jsonify({"message": "Data saved successfully"}), 200
 
 
 if __name__ == '__main__':
