+++ cite_envs/sample.yml
@@ -1,9 +1,9 @@
-lastupdate: '2023-09-09 21:17:09.906044'
+lastupdate: '2023-09-30 14:23:57.318304'
 nextpage: ''
 position:
   money:
     X: 169
-    Y: 484
+    Y: 200
     h: -18
     w: -166
   star:
