+++ util/ocr_path.txt
@@ -1,1 +1,1 @@
-D:\program\Tesseract-OCR\tesseract.exe
+D:\python-training\ocr_crawler\Tesseract-OCR\tesseract.exe
